phpstan rule level increased to 3
Updates #411

+++ crm/src/Application/Models/Ticket.php
@@ -796,7 +796,7 @@ class Ticket extends ActiveRecord
 				: new \DateTime($this->getClosedDate());
 			$diff = $targetDate->diff($dateEntered);
 			$daysPassed = $diff->format('%a');
-			return round($daysPassed/$days*100);
+			return (int)round($daysPassed/$days*100);
 		}
 		return null;
 	}
